putting GridGapControlComponent into a component to fix HMR
