+++ editor/src/components/canvas/canvas-strategies/strategies/set-grid-gap-strategy.tsx
@@ -40,7 +40,7 @@ import {
 import type { InteractionSession } from '../interaction-state'
 import { colorTheme } from '../../../../uuiui'
 import { activeFrameTargetPath, setActiveFrames } from '../../commands/set-active-frames-command'
-import type { GridGapControlProps } from '../../controls/select-mode/grid-gap-control'
+import type { GridGapControlProps } from '../../controls/select-mode/grid-gap-control-component'
 import { GridGapControl } from '../../controls/select-mode/grid-gap-control'
 import type { GridControlsProps } from '../../controls/grid-controls-for-strategies'
 import { controlsForGridPlaceholders } from '../../controls/grid-controls-for-strategies'
